test: separate package install smoke coverage
